Move validator into domain
